cookbook/python/streamlit-chatbot/requirements.txt
streamlit
weaviate-client
openai
surfer-protocol